fix(cloud-templates): cast default number and boolean numbers to feel
related to https://github.com/camunda/camunda-modeler/issues/4517
