+++ src/cloud-element-templates/Helper.js
@@ -3,6 +3,7 @@ import { getBusinessObject } from 'bpmn-js/lib/util/ModelUtil';
 import { is, isAny } from 'bpmn-js/lib/util/ModelUtil';
 
 import { v4 as uuid } from 'uuid';
+import { isSpecialFeelProperty, toFeelExpression } from './util/feelUtil';
 
 /**
  * The BPMN 2.0 extension attribute name under
@@ -176,6 +177,12 @@ export function findZeebeSubscription(message) {
 }
 
 export function getDefaultValue(property) {
+
+  if (
+    isSpecialFeelProperty(property)
+  ) {
+    return toFeelExpression(property.value, property.type);
+  }
 
   if (property.value !== undefined) {
     return property.value;
