+++ src/cloud-element-templates/properties-panel/properties/custom-properties/NumberProperty.js
@@ -1,6 +1,7 @@
 import { FeelNumberEntry, NumberFieldEntry } from '@bpmn-io/properties-panel';
 import { useService } from 'bpmn-js-properties-panel';
-import { isSpecialFeelProperty, propertyValidator, usePropertyAccessors } from './util';
+import { propertyValidator, usePropertyAccessors } from './util';
+import { isSpecialFeelProperty } from '../../../util/feelUtil';
 import { PropertyDescription } from '../../../../components/PropertyDescription';
 import { PropertyTooltip } from '../../components/PropertyTooltip';
 import { useCallback } from '@bpmn-io/properties-panel/preact/hooks';
